CI Framework
Removes the dependency on nfdump and perform this ourselves this allows beaker
to use 'puppet module install' exclusively for all prerequisite modules.  The
installation code has been made generic in that vanilla sources will be
installed by default, with the option of using a private vcs repository if
required.  Adds in basic beaker acceptance testing of Ubuntu 14.04.